Refactor phone radio server
Move Vonage client setup into phone-radio/src/clients.ts.

Move route handlers into flat phone-radio/src/routes.ts.

Move track list into phone-radio/src/tracks.ts.

Staged files: phone-radio/src/clients.ts, phone-radio/src/index.ts, phone-radio/src/routes.ts, phone-radio/src/tracks.ts.
